extract list sessions
